separate dataNodes management and dataNodes filtering
